+++ CHANGELOG.md
@@ -72,6 +72,7 @@ All user visible changes to this project will be documented in this file. This p
     - Chat info page:
         - Redesigned chat info page. ([#265], [#256])
     - Redesigned scrollbar. ([#276], [#262])
+    - Redesigned snackbar. ([#336])
 
 ### Fixed
 
@@ -165,6 +166,7 @@ All user visible changes to this project will be documented in this file. This p
 [#323]: /../../pull/323
 [#328]: /../../issues/328
 [#332]: /../../pull/332
+[#336]: /../../pull/336
 [#337]: /../../pull/337